ebook_library_sync/requirements.txt
# Dependências para Ebook Library Sync
# Python >= 3.7

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Opcional para análise de dados
pandas>=1.5.0

# Opcional para melhor logging
colorlog>=6.7.0
